- DirectXInput: Started adding gyro DSU client support. (WIP2) - DirectXInput: Added more accurate system tick checks.

+++ FpsOverlayer/AppVariables.cs
@@ -27,7 +27,7 @@ namespace FpsOverlayer
         public static int vKeypadBottomMargin = 0;
 
         //Frames per second
-        public static int vLastFrameTimeAdded = 0;
+        public static long vLastFrameTimeAdded = 0;
         public static double vLastFrameTimeStamp = 0;
         public static List<double> vListFrameTime = new List<double>();
