+++ LibraryShared/Classes/ControllerButtonDetails.cs
@@ -11,8 +11,8 @@ namespace LibraryShared
             public bool PressedShort { get; set; } = false;
             public bool PressedLong { get; set; } = false;
             public bool PressTimeDone { get; set; } = false;
-            public int PressTimeCurrent { get; set; } = 0;
-            public int PressTimePrevious { get; set; } = 0;
+            public long PressTimeCurrent { get; set; } = 0;
+            public long PressTimePrevious { get; set; } = 0;
         }
     }
 }
